move cuckoo structs to protocol folder
